refac(spec): evaluator in its own module; coverage checked in one tree walk
evaluate.py holds the recursive evaluator, builder.py the declarations.
check_coverage collects divisor, constant-side and coefficient obligations
in one walk, so cases: masks are evaluated once per declaration.
Public docstrings in numpy style so the API pages render.
