+++ linopy/spec/coverage.py
@@ -12,7 +12,8 @@ answered.
 
 from __future__ import annotations
 
-from collections.abc import Iterator
+from collections.abc import Sequence
+from dataclasses import dataclass, field
 
 import xarray as xr
 from math_spec import program as ms
@@ -24,6 +25,16 @@ from linopy.spec.nodes import amounts_of, children, parameters_of
 from linopy.spec.where import evaluate_where
 
 Rows = xr.DataArray | None
+Obligation = tuple[str, Rows]
+
+
+@dataclass
+class Obligations:
+    """Every parameter use under a declaration, each with the rows it has to cover, gathered in one walk."""
+
+    divisors: list[Obligation] = field(default_factory=list)
+    constants: list[Obligation] = field(default_factory=list)
+    coefficients: list[Obligation] = field(default_factory=list)
 
 
 def gaps_under(array: xr.DataArray, rows: Rows) -> int:
@@ -32,6 +43,133 @@ def gaps_under(array: xr.DataArray, rows: Rows) -> int:
     if rows is not None:
         missing = missing & rows
     return int(missing.sum())
+
+
+def check_coverage(
+    subject: str,
+    expressions: Sequence[ms.ExpressionNode],
+    ctx: Context,
+    rows: Rows,
+    *,
+    comparison: bool = False,
+) -> None:
+    """
+    Refuse *subject* if a parameter it reads leaves a row it builds uncovered.
+
+    One walk over *expressions* collects what every parameter has to cover,
+    narrowed at each ``cases:`` region; divisors are judged first, then, for a
+    *comparison*, the side without a variable term, then every coefficient.
+    """
+    found = obligations_of(expressions, ctx, rows, comparison=comparison)
+    check_divisors(subject, found.divisors, ctx)
+    check_constant_sides(subject, found.constants, ctx)
+    check_coefficients(subject, found.coefficients, ctx)
+
+
+def obligations_of(
+    expressions: Sequence[ms.ExpressionNode],
+    ctx: Context,
+    rows: Rows,
+    *,
+    comparison: bool = False,
+) -> Obligations:
+    """What the parameters under *expressions* have to cover, a side of a *comparison* without a variable being its constant side."""
+    found = Obligations()
+    for expression in expressions:
+        constant = comparison and not ms.carries_variable(expression)
+        _collect(expression, ctx, rows, constant, found)
+    return found
+
+
+def _collect(
+    node: ms.ExpressionNode,
+    ctx: Context,
+    rows: Rows,
+    constant: bool,
+    into: Obligations,
+) -> None:
+    if isinstance(node, ms.Divide):
+        into.divisors.extend(_divisor_uses(node, ctx, rows))
+    if isinstance(node, ms.Parameter):
+        if constant:
+            into.constants.append((node.name, rows))
+        into.coefficients.append((node.name, rows))
+    into.coefficients.extend((name, None) for name in amounts_of(node))
+    if isinstance(node, ms.Cases):
+        for region in node.regions:
+            inside = evaluate_where(region.when, ctx)
+            narrowed = inside if rows is None else rows & inside
+            _collect(region.value, ctx, narrowed, constant, into)
+        return
+    for child in children(node):
+        _collect(child, ctx, rows, constant, into)
+
+
+def _divisor_uses(quotient: ms.Divide, ctx: Context, rows: Rows) -> list[Obligation]:
+    """Each parameter in the divisor, with the rows the quotient is divided over: the region, narrowed by the presence of every numerator variable."""
+    params = parameters_of(quotient.divisor)
+    if not params:
+        return []
+    needed = rows
+    for variable in sorted(ms.variables_of(quotient.numerator)):
+        present = terms.present(ctx.model.variables[variable])
+        needed = present if needed is None else needed & present
+    return [(param, needed) for param in sorted(params)]
+
+
+def check_divisors(subject: str, found: Sequence[Obligation], ctx: Context) -> None:
+    """
+    A divisor must have a value wherever *subject* divides by it.
+
+    Reached before evaluation, the last moment the gap is visible: the
+    coefficient fill would turn it into a division by zero.
+    """
+    for param, needed in found:
+        missing = gaps_under(ctx.parameters[param], needed)
+        if missing:
+            raise SpecDataError(
+                f"{subject}: parameter '{param}' is used as a divisor but covers {missing} "
+                f"fewer coordinates than it is divided over. A missing row means a zero "
+                f"coefficient everywhere else, and zero is not a divisor: the term would drop "
+                f"and the row would silently stop constraining.\n"
+                f"  Supply the missing rows, or mask the coordinates out with a where."
+            )
+
+
+def check_constant_sides(
+    subject: str, found: Sequence[Obligation], ctx: Context
+) -> None:
+    """A comparison's constant side must have values wherever the row is built, or the zero is the bound."""
+    for param, needed in sorted(found, key=lambda pair: pair[0]):
+        missing = gaps_under(ctx.parameters[param], needed)
+        if missing:
+            raise SpecDataError(
+                f"{subject}: parameter '{param}' covers {missing} fewer coordinates "
+                f"than the rows built here. A missing row is read as 0, and on the constant side "
+                f"that zero is a bound rather than an absence: the row still exists, and it binds.\n"
+                f"  Supply the missing rows, if the value is what was meant.\n"
+                f"  Mask them out with a where, if the row should not exist there."
+            )
+
+
+def check_coefficients(subject: str, found: Sequence[Obligation], ctx: Context) -> None:
+    """
+    A coefficient parameter must reach every row it is built over.
+
+    A missing coefficient row would otherwise read as a zero, dropping its term
+    while the row stays. A shift offset or window width given by name is a
+    coefficient too, and stands or falls over its own coordinates.
+    """
+    for param, needed in found:
+        missing = gaps_under(ctx.parameters[param], needed)
+        if missing:
+            raise SpecDataError(
+                f"{subject}: parameter '{param}' is used as a coefficient but leaves "
+                f"{missing} of the rows built here uncovered. A missing row reads as a zero "
+                f"coefficient, dropping the term while the row stays.\n"
+                f"  Supply the missing rows, if a value other than 0 was meant.\n"
+                f"  Mask them out with a where, if the row should not exist there."
+            )
 
 
 def check_bounds_cover(
@@ -48,118 +186,3 @@ def check_bounds_cover(
             f"  supply the value           the variable exists there, bounded (`inf` is a value)\n"
             f'  where: "<the parameter>"   the variable does not exist there at all'
         )
-
-
-def check_constant_side_covers(
-    name: str, row: ms.ConstraintDeclaration, ctx: Context, rows: Rows
-) -> None:
-    """A comparison's constant side must have values wherever the row is built, or the zero is the bound."""
-    for side in (row.lhs, row.rhs):
-        if ms.carries_variable(side):
-            continue
-        found = sorted(
-            (
-                (node.name, narrowed)
-                for node, narrowed in _under_regions(side, ctx, rows)
-                if isinstance(node, ms.Parameter)
-            ),
-            key=lambda pair: pair[0],
-        )
-        for param, narrowed in found:
-            missing = gaps_under(ctx.parameters[param], narrowed)
-            if missing:
-                raise SpecDataError(
-                    f"constraint '{name}': parameter '{param}' covers {missing} fewer coordinates "
-                    f"than the rows built here. A missing row is read as 0, and on the constant side "
-                    f"that zero is a bound rather than an absence: the row still exists, and it binds.\n"
-                    f"  Supply the missing rows, if the value is what was meant.\n"
-                    f"  Mask them out with a where, if the row should not exist there."
-                )
-
-
-def check_divisors_cover(
-    subject: str, expressions: tuple[ms.ExpressionNode, ...], ctx: Context, rows: Rows
-) -> None:
-    """
-    A divisor must have a value wherever *subject* divides by it.
-
-    The rows that ask are the declaration's own, narrowed by the presence of
-    every variable in the quotient's numerator and by the region of a
-    ``cases:`` block. Reached before evaluation, the last moment the gap is
-    visible: the coefficient fill would turn it into a division by zero.
-    """
-    for expression in expressions:
-        for quotient, region in _under_regions(expression, ctx, rows):
-            if not isinstance(quotient, ms.Divide):
-                continue
-            params = parameters_of(quotient.divisor)
-            if not params:
-                continue
-            needed = region
-            for variable in sorted(ms.variables_of(quotient.numerator)):
-                present = terms.present(ctx.model.variables[variable])
-                needed = present if needed is None else needed & present
-            for param in sorted(params):
-                missing = gaps_under(ctx.parameters[param], needed)
-                if missing:
-                    raise SpecDataError(
-                        f"{subject}: parameter '{param}' is used as a divisor but covers {missing} "
-                        f"fewer coordinates than it is divided over. A missing row means a zero "
-                        f"coefficient everywhere else, and zero is not a divisor: the term would drop "
-                        f"and the row would silently stop constraining.\n"
-                        f"  Supply the missing rows, or mask the coordinates out with a where."
-                    )
-
-
-def check_coefficients_cover(
-    subject: str, expressions: tuple[ms.ExpressionNode, ...], ctx: Context, rows: Rows
-) -> None:
-    """
-    A coefficient parameter must reach every row it is built over.
-
-    A missing coefficient row would otherwise read as a zero, dropping its term
-    while the row stays. Decided against the rows the declaration builds,
-    narrowed at each ``cases:`` region exactly as the other checks are, so a
-    ``where`` that removed the coordinate has already answered. A shift offset
-    or window width given by name is a coefficient too, and stands or falls
-    over its own coordinates.
-    """
-    for expression in expressions:
-        for node, region in _under_regions(expression, ctx, rows):
-            for param, needed in _coefficient_uses(node, region):
-                missing = gaps_under(ctx.parameters[param], needed)
-                if missing:
-                    raise SpecDataError(
-                        f"{subject}: parameter '{param}' is used as a coefficient but leaves "
-                        f"{missing} of the rows built here uncovered. A missing row reads as a zero "
-                        f"coefficient, dropping the term while the row stays.\n"
-                        f"  Supply the missing rows, if a value other than 0 was meant.\n"
-                        f"  Mask them out with a where, if the row should not exist there."
-                    )
-
-
-def _coefficient_uses(
-    node: ms.ExpressionNode, region: Rows
-) -> Iterator[tuple[str, Rows]]:
-    """Each parameter *node* uses as a coefficient, with the rows it has to cover."""
-    if isinstance(node, ms.Parameter):
-        yield node.name, region
-        return
-    for name in amounts_of(node):
-        yield name, None
-
-
-def _under_regions(
-    node: ms.ExpressionNode, ctx: Context, rows: Rows
-) -> Iterator[tuple[ms.ExpressionNode, Rows]]:
-    """Every node under *node* with the rows it has to cover, narrowed at each ``cases:`` region."""
-    yield node, rows
-    if isinstance(node, ms.Cases):
-        for region in node.regions:
-            inside = evaluate_where(region.when, ctx)
-            yield from _under_regions(
-                region.value, ctx, inside if rows is None else rows & inside
-            )
-        return
-    for child in children(node):
-        yield from _under_regions(child, ctx, rows)
